requirements.txt
flask==2.0.3
transformers==4.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==1.5.3
numpy==1.23.5
requests==2.28.2
beautifulsoup4==4.12.2
torch==2.0.1
python-dotenv==1.0.0
tqdm==4.65.0